requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
crispy-bootstrap5==0.6
Django==4.0.4
django-crispy-forms==1.14.0
idna==3.3
Pillow==9.1.1
psycopg2-binary==2.9.3
requests==2.27.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
stripe==3.1.0
urllib3==1.26.9
